Deprecate support for building AUR packages
There are some potential security issues with this feature, or rather in
how it is implemented. The main problem is the ultimate trust in `yay`
and the lack of verification of the downloaded archive.

To be more specific, because the download always contains the latest
snapshot, i.e. not a static version, and is unverified a malicous change
could easliy be included without any warning. This could happen if the
aur package maintainer's account was compromised, or abandons the
maintainership.

There are partial solutions to this problem but it would mean more work
to maintain this image. This feature is currently outside the scope of
this derivative project, so it is removed.