Remove AUR support
This is mainly removed because it won't be used. There is however also
the potential security and trust issue to be considered.

The current AUR support relies on building and installing an untrusted
package (aur helper) by default. Since it will be part of the image it
will affect everyone, with no ability to opt out.

AUR support could be re-added by, for example, only relying upon
makepkg. The important part is that untrusted packages should not be
installed by default and certainly not in an insecure manner.