Change base to archimg/base-devel
Although there is some conflicting information on the Docker Hub, it
would seem base/archlinux is actually not much more than an alias now
for archimg/base.

In any case the following repo is now the source:

- https://github.com/archimg/archlinux

Advantages that come with this change is that base-devel no longer needs
to be installed if archimg/base-devel is used.

There also should be no need for using reflector. Although archimg/base
does not do anything similar I trust the maintainers of that image with
that responsibility. Personally I did have some issues with package repos
not responding due to this.