Use a more secure sudoers file
The sudoers file granted NOPASSWD for all commands, when the travis user
only really needs access to `pacman`.

It does need it for `sed` as well with arch-travis.sh, but that could
technically be replaced by some ownership and permission changes and
an #include in pacman.conf.

Using `groupadd` is redundant since, by default a matching group
is added by userdel.