Fix meson crash due to $CC being null
Because the script `arch-travis.sh` always sets the $CC environment
variable, it will either contain a specified compiler or contain a null
value. It will never be unset.

Meson, the build system, does not handle this resulting in a crash.

Ideally it should never be set, unless it is needed. But it proved
simpler to just explicitly set CC=gcc by default.