Remove reflector
Not sure how, but it seems to cause some issues with mirrors not
responding. Despite having 10 mirrors configured `pacman` sometimes get
stuck at the first one and does not timeout.

Not sure about this one. This might all be wrong and in that case it is
simple to revert this.